# openapi/api-gateway.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/gateway/tenants": {
      "post": {
        "operationId": "TenantGatewayController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTenantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "TenantGateway"
        ]
      },
      "get": {
        "operationId": "TenantGatewayController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "TenantGateway"
        ]
      }
    },
    "/gateway/tenants/{id}": {
      "get": {
        "operationId": "TenantGatewayController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "TenantGateway"
        ]
      },
      "patch": {
        "operationId": "TenantGatewayController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTenantDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "TenantGateway"
        ]
      },
      "delete": {
        "operationId": "TenantGatewayController_softDelete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "TenantGateway"
        ]
      }
    },
    "/gateway/tenants/by-name/{name}": {
      "get": {
        "operationId": "TenantGatewayController_findByName",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "TenantGateway"
        ]
      }
    },
    "/gateway/tenants/{id}/status": {
      "patch": {
        "operationId": "TenantGatewayController_changeStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "TenantGateway"
        ]
      }
    },
    "/gateway/tenants/by-name/{name}/status": {
      "patch": {
        "operationId": "TenantGatewayController_changeStatusByName",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "TenantGateway"
        ]
      },
      "get": {
        "operationId": "TenantGatewayController_getStatusByName",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "TenantGateway"
        ]
      }
    },
    "/gateway/tenants/validate/{name}": {
      "get": {
        "operationId": "TenantGatewayController_validateTenant",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "TenantGateway"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Gateway",
    "description": "Public HTTP entrypoint routing to microservices (TCP).",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://aero.darmist.com/backend",
      "description": "Production Server"
    },
    {
      "url": "http://localhost:3501",
      "description": "Local Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Use a valid **access token**.\n\nFormat: `Bearer <ACCESS_TOKEN>`",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateTenantDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateTenantDto": {
        "type": "object",
        "properties": {}
      },
      "ChangeStatusDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}